+++ app/src/main/java/com/orange/ods/app/ui/components/ComponentDemoScreen.kt
@@ -39,6 +39,7 @@ fun ComponentDemoScreen(componentId: Long) {
             Component.Dialogs -> ComponentDialog()
             Component.FloatingActionButtons -> ComponentFloatingActionButton()
             Component.Lists -> ComponentLists()
+            Component.ModalDrawers -> ComponentModalDrawers()
             Component.RadioButtons -> ComponentRadioButtons()
             Component.SheetsBottom -> ComponentSheetsBottom()
             Component.Sliders -> ComponentSliders()